Delete other elements
If delete user account, remove related tournaments, matches, team

+++ back/app/crud/match.py
@@ -7,6 +7,7 @@ from fastapi import HTTPException
 from app.models.match import Match
 from app.models.action import Action
 from app.schemas.action import ActionTypeEnum
+from app.crud.tournament import check_and_update_tournament_status
 
 
 def create_match(db: Session, match: MatchCreate):
@@ -63,8 +64,14 @@ def update_match(db: Session, match: MatchUpdate, match_id: int):
 
 def delete_match(db: Session, match_id: int):
     db_match = get_match(db=db, match_id=match_id)
-    db.delete(db_match)
-    db.commit()
+    if db_match:
+        # Delete all actions related to the match
+        db.query(Action).filter(Action.match_id == match_id).delete()
+        db.commit()
+
+        # Delete the match
+        db.delete(db_match)
+        db.commit()
     return db_match
 
 def get_match_player(db:Session, match_id:int):
@@ -79,26 +86,40 @@ def get_match_player(db:Session, match_id:int):
     }
 
 def calculate_match_result(db: Session, match_id: int):
+    # Pobranie wszystkich akcji z meczu
     actions = db.query(Action).filter(Action.match_id == match_id).all()
-    team_goals = {}
 
+    # Sprawdzenie, czy istnieją jakiekolwiek akcje
+    if not actions:
+        return None  # Brak akcji, nic do przeliczenia
+
+    # Pobranie informacji o meczu
+    db_match = db.query(Match).filter(Match.id == match_id).first()
+    if not db_match:
+        return None  # Mecz nie istnieje
+
+    # Identyfikatory drużyn z meczu
+    team_1_id = db_match.team_1_id
+    team_2_id = db_match.team_2_id
+
+    # Zliczanie goli dla każdej drużyny
+    team_goals = {team_1_id: 0, team_2_id: 0}
     for action in actions:
         if action.action_type == ActionTypeEnum.Goal:
-            if action.team_id not in team_goals:
-                team_goals[action.team_id] = 0
-            team_goals[action.team_id] += 1
+            if action.team_id in team_goals:
+                team_goals[action.team_id] += 1
 
-    if len(team_goals) == 2:
-        team_ids = list(team_goals.keys())
-        result = f"{team_goals[team_ids[0]]}:{team_goals[team_ids[1]]}"
-    else:
-        result = None
+    # Przygotowanie wyniku w formacie "X:Y"
+    result = f"{team_goals.get(team_1_id, 0)}:{team_goals.get(team_2_id, 0)}"
 
-    db_match = db.query(Match).filter(Match.id == match_id).first()
-    if db_match:
-        db_match.result = result
-        db.commit()
-        db.refresh(db_match)
+    # Aktualizacja wyniku w tabeli Match
+    db_match.result = result
+    db.commit()
+    db.refresh(db_match)
+
+    # Aktualizacja statusu turnieju, jeśli dotyczy
+    check_and_update_tournament_status(db, db_match.tournament_id)
 
     return db_match
+
